requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic-settings==2.5.2
pyyaml==6.0.2
anthropic==0.34.2
tenacity==9.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
